+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Separate global and iterative switching algorithm objective functions
 - Updated to use Graphs instead of LightGraphs (LightGraphs was archived)
 - Fixed bug in Distributed usage, where running on multiple workers but not intending to run distributed algorithm
 - Added Distributed computing options for fault studies and stability analysis to significantly speed up these processes